fix(csa-executor): align codex CLI session argv
Use `codex exec --json` for new sessions and switch resume calls to
`codex exec resume <session> [prompt]`, preserving stdin transport by
passing `-` when a resumed prompt is piped.

### AI Reviewer Metadata
- **Design Intent**: Fix the default codex CLI transport so provider session IDs are captured and reused for resume and native fork flows.
- **Key Decisions**: Added `--json` on the persistent codex CLI path, rewrote resume argv to the positional `resume` subcommand, and split prompt helpers plus codex resume tests into focused files to satisfy the monolith gate.
- **Reviewer Guidance**: Focus review on non-ACP codex persistent sessions, especially provider session extraction from JSONL, resumed long-prompt stdin behavior, and the no-op file splits required by pre-commit.
